pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "whisper-transcribe"
version = "0.1.0"
authors = [
    { name = "Leon Melamud" }
]
description = "A PyQt6-based GUI application for audio transcription using Whisper"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "transformers>=4.37.2",
    "datasets>=2.16.1",
    "numpy>=1.24.3",
    "ffmpeg-python>=0.2.0",
    "python-dotenv>=1.0.0",
    "PyQt6>=6.4.0",
    "tenacity>=8.2.0",
    "typing-extensions>=4.5.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["whisper_transcribe"]
